probability of apples

+++ src/org/academiadecodigo/macasdoze/AppleCollector.java
@@ -23,12 +23,13 @@ public class AppleCollector {
         if (appleList.size() != 0) {
 
             for (int i = 0; i < appleList.size(); i++) {
-                if (appleList.get(i).getPosition().getY() > newton.getPosition().getY() && (appleList.get(i).getPosition().getX() < newton.getPosition().getX() + newton.getPosition().getWidth()
+                if ((appleList.get(i).getPosition().getY() > newton.getPosition().getY()
+                        && appleList.get(i).getPosition().getY() + appleList.get(i).getPosition().getHeight() < newton.getPosition().getY() + newton.getPosition().getHeight())
+                        && (appleList.get(i).getPosition().getX() < newton.getPosition().getX() + newton.getPosition().getWidth()
                         && appleList.get(i).getPosition().getWidth() + appleList.get(i).getPosition().getX() > newton.getPosition().getX())) {
-                    newton.increasePoints(appleList.get(i).getPoints());
+                    Score.increaseScore(appleList.get(i).getPoints());
                     appleList.get(i).getPosition().deleteObject();
                     appleList.remove(i);
-                    System.out.println(newton.getPoints());
                 }
             }
         }
